Add files via upload
CRICKET CHATBOT
Introduction
The name of the chatbot is “MEERA”, developed in python. This chatbot will tell you all the information of a cricket match. The chatbot uses various APIs to process the information.

Features
The chatbot tells the following details about a cricket match-
•	Description of all ongoing live matches at any particular time.
•	Description includes series name, team names, date of match.
•	Live score of a match played between two teams.
•	Run rate 
•	Name of the ground(venue), where match is played.
•	Current status of the match
•	Result of the match
•	Format of the match (T20, Test, ODI)



The algorithm uses google API for speech recognition. For fetching information of a cricket match, algorithm uses espncricinfo API.

Requirements 
Following are the list of python libraries that need to be installed.
•	BS4
•	Requests
•	Espncricinfo
•	Gtts
•	Os
•	Speech_recognition



Limitations
The chatbot will not give info about any domestic match. It will cover only international matches.